+++ src/hera/__init__.py
@@ -72,3 +72,7 @@ from hera.workflow_service import WorkflowService
 from hera.workflow_status import WorkflowStatus
 from hera.workflow_template import WorkflowTemplate
 from hera.workflow_template_service import WorkflowTemplateService
+from hera._version import version
+
+__version__ = version
+__version_info__ = version.split('.')