+++ CHANGELOG.md
@@ -35,6 +35,10 @@ The general format is:
 
 - tolerations can now be set via workflow
 
+### Changed
+
+- Fix float type handling for max_cpu and min_cpu properties in Resources class. 
+
 # 3.6.4 - DATE (11/08/2022)
 
 ### Removed