+++ CHANGELOG.md
@@ -37,7 +37,7 @@ The general format is:
 
 ### Changed
 
-- Fix float type handling for max_cpu and min_cpu properties in Resources class. 
+- float type handling for `max_cpu` and `min_cpu` properties in `Resources` class
 
 # 3.6.4 - DATE (11/08/2022)