+++ CHANGELOG.md
@@ -36,6 +36,7 @@ The general format is:
 - support for Git artifact authentication credentials
 - tolerations can now be set via workflow
 - version via `hera.__version__`
+- volume specifications on workflows
 
 ### Changed